Style improvement: mirror functions, less #define tricks, independent execution
QFlex tricks in order to ensure single stepping where made by injecting code
through #defines. This makes code very hard to read and also merging with main
QEMU difficult because of inter-dependencies. By making mirror functions of QEMU
in the places we modify execution flow, it increases readability and
upgradability for the future.