QFLEX: qflex_cpu_exec as in cpu_exec (no break of loop)
Observed beheaviour of breaking in the current point traps QEMU into a infinite
interrupt queue. This comes from nested interrupts, as the return address is an
interrupt itself, it keeps looping on ERET, never going back to user mode.

FLAGS activation and checks for single stepping by breaking the loop are still
in  similar spots as defined in FLEXUS, but not used.