FEATURE: profiling, extract statistic of instruction types
If enabled, it generated helpers that gather statistic of instruction types
correspoding to the decode tree.
Example:
  some System instruction will add a count to
  L1H in B_EXEC_SYS (Branch, exception generation and system insns)
  L2H in BRES_SYSTEM

This information is available in qflex_insn_profile_l1h and
qflex_insn_profile_l2h.
The system also makes a difference when an instruction is exectued in Kernel or
User mode.

Note: qflex.profile_enable make it such as QEMU generates the HELPER's, but in order
to activate the statistic gathering you must enable qflex.profiling (see helper
HELPER(qflex_profile)). This could/should be done with a magic instruction.