Fix tcg tests SKIPPING by adding cross-compilation depedencies
As documented in multiple locations describing what check-tcg does, it is actually a requirement to have the cross compiling tools pre-installed in order compile and run the test files.
I've fixed the SKIPPED tests by adding cross compilation tools for aarch64 in the list of dependencies `scripts/qflex/install-deps.sh`.