Remove TCI test
I though originally this was the only tests to launch boot-serial and CD-ROM tests, and thus I enabled it.
Given that my understanding was wrong, TCI is actually a very niche feature that doesn't seem to really been maintained since 2011 and with little documentation.
From my understanding TCI is there to execute TCG binary code generated by QEMU, similarly to Java, TCI would stand as the JVM and Java Code as an intermediate representation of TCG binary.
It is useful to support very nieche host platforms such as those not listed here (arm, i386, ia64, ppc, ppc64, s390, sparc, x86_64).
We do not require this test.